+++ xds/src/main/java/io/grpc/xds/XdsEndpointResource.java
@@ -30,8 +30,8 @@ import io.grpc.EquivalentAddressGroup;
 import io.grpc.xds.ClientXdsClient.ResourceInvalidException;
 import io.grpc.xds.Endpoints.DropOverload;
 import io.grpc.xds.Endpoints.LocalityLbEndpoints;
-import io.grpc.xds.XdsEndpointResource.EdsUpdate;
 import io.grpc.xds.XdsClient.ResourceUpdate;
+import io.grpc.xds.XdsEndpointResource.EdsUpdate;
 import java.net.InetSocketAddress;
 import java.util.ArrayList;
 import java.util.Collections;
